Handle unexpected exits
One common way to indicate failure of a process it to exit non-zero.
Prior to these changes, littlecheck did not care about the exit code,
aside from showing information about signal termination if other checks
did not pass.

This commit changes littlecheck's behavior in the following ways:
- CHECKERR is no longer combined with CHECK lines if both report errors.
  This simplifies checking of the two streams and makes results more
  consistent.
- It is now possible to specify how the process is expected to exit via
  CHECKEXIT lines. Each line can specify an arbitrary number of
  whitespace-separated exit conditions, and there can be an arbitrary
  number of CHECKEXIT lines.
  If one of the specified exit conditions matches how the process
  actually exits, it is considered an expected exit. Otherwise, the test
  will fail and inform about the unexpected exit condition.
  If no exit condition is specified, the process is expected to exit
  with code 0.
  An exit condition can be an integer, with non-negative integers being
  interpreted as the literal exit code, and negative integers are
  interpreted to indicate that the process is expected to exit with the
  negated signal value (e.g. specifying `-9` indicates that the process
  is expected to exit with signal 9, a.k.a. SIGTERM). Alternatively,
  signal names can be specified in uppercase, either with the `SIG`
  prefix (e.g. SIGTERM) or without it (e.g. TERM).

This is a breaking change for users who have tests which exit nonzero in
cases which should be considered successful tests. The tests will have
to be modified to either always exit 0 on success, or by adding an
annotation specifying the expected exit codes, e.g.
`# CHECKEXIT: 1`